+++ apps/rosen-service/src/scanner/scanner-service.ts
@@ -1,10 +1,8 @@
 import WinstonLogger from '@rosen-bridge/winston-logger/dist/WinstonLogger';
 
+import { startBitcoinScanner } from './chains/bitcoin';
 import { startCardanoScanner } from './chains/cardano';
 import { startErgoScanner } from './chains/ergo';
-
-import observationService from '../observation/observation-service';
-import eventTriggerService from '../event-trigger/event-trigger-service';
 
 import { handleError } from '../utils';
 
@@ -15,13 +13,18 @@ const logger = WinstonLogger.getInstance().getLogger(import.meta.url);
  */
 const start = async () => {
   try {
-    const [ergoScanner, cardanoScanner] = await Promise.all([
+    const [ergoScanner, cardanoScanner, bitcoinScanner] = await Promise.all([
       startErgoScanner(),
       startCardanoScanner(),
+      startBitcoinScanner(),
     ]);
 
     logger.debug('all scanners started and their extractors registered', {
-      scannerNames: [ergoScanner.name(), cardanoScanner.name()],
+      scannerNames: [
+        ergoScanner.name(),
+        cardanoScanner.name(),
+        bitcoinScanner.name(),
+      ],
     });
   } catch (error) {
     handleError(error);