+++ wallets/eternl-wallet/src/eternlWalletInfo.ts
@@ -1,0 +1,9 @@
+import { EternlIcon } from '@rosen-bridge/icons';
+import { WalletInfo } from '@rosen-ui/wallet-api';
+
+export const eternlWalletInfo: WalletInfo = {
+  icon: EternlIcon,
+  name: 'Eternl',
+  label: 'Eternl',
+  link: 'https://eternl.io/',
+};