+++ wallets/nami-wallet/src/namiWalletCreator.ts
@@ -1,0 +1,17 @@
+import { Wallet } from '@rosen-ui/wallet-api';
+
+import { getBalanceCreator } from './getBalance';
+import { getNamiWallet } from './getNamiWallet';
+import { isNamiAvailable } from './isNamiAvailable';
+import { transferCreator } from './transfer';
+import { NamiWalletCreator } from './types';
+
+export const namiWalletCreator = (
+  config: NamiWalletCreator
+): Wallet | undefined => {
+  if (!isNamiAvailable()) return;
+  return Object.assign({}, getNamiWallet(), {
+    getBalance: getBalanceCreator(config),
+    transfer: transferCreator(config),
+  });
+};