+++ wallets/nami-wallet/src/namiWalletInfo.ts
@@ -1,0 +1,9 @@
+import { NamiIcon } from '@rosen-bridge/icons';
+import { WalletInfo } from '@rosen-ui/wallet-api';
+
+export const namiWalletInfo: WalletInfo = {
+  icon: NamiIcon,
+  name: 'Nami',
+  label: 'Nami',
+  link: 'https://namiwallet.io/',
+};