+++ wallets/nautilus-wallet/src/transfer.ts
@@ -1,0 +1,50 @@
+import { RosenChainToken } from '@rosen-bridge/tokens';
+import { convertNumberToBigint, validateDecimalPlaces } from '@rosen-ui/utils';
+
+import { getNautilusWallet } from './getNautilusWallet';
+import { NautilusWalletCreator } from './types';
+
+export const transferCreator =
+  (config: NautilusWalletCreator) =>
+  async (
+    token: RosenChainToken,
+    decimalAmount: number,
+    toChain: string,
+    toAddress: string,
+    decimalBridgeFee: number,
+    decimalNetworkFee: number,
+    lockAddress: string
+  ): Promise<string> => {
+    validateDecimalPlaces(decimalAmount, token.decimals);
+    validateDecimalPlaces(decimalBridgeFee, token.decimals);
+    validateDecimalPlaces(decimalNetworkFee, token.decimals);
+
+    const wallet = await getNautilusWallet().api.getContext();
+    const tokenId = token.tokenId;
+    const amount = convertNumberToBigint(decimalAmount * 10 ** token.decimals);
+    const bridgeFee = convertNumberToBigint(
+      decimalBridgeFee * 10 ** token.decimals
+    );
+    const networkFee = convertNumberToBigint(
+      decimalNetworkFee * 10 ** token.decimals
+    );
+    const changeAddress = await wallet.get_change_address();
+
+    const walletUtxos = await wallet.get_utxos();
+    if (!walletUtxos) throw Error(`No box found`);
+
+    const unsignedTx = await config.generateUnsignedTx(
+      changeAddress,
+      walletUtxos,
+      lockAddress,
+      toChain,
+      toAddress,
+      tokenId,
+      amount.toString(),
+      bridgeFee.toString(),
+      networkFee.toString()
+    );
+    const signedTx = await wallet.sign_tx(unsignedTx);
+    const result = await wallet.submit_tx(signedTx);
+    return result;
+  };