+++ wallets/nautilus-wallet/src/utils.ts
@@ -1,7 +1,5 @@
-'use server';
-
 import * as wasm from 'ergo-lib-wasm-nodejs';
-import { minBoxValue } from './consts';
+import { minBoxValue } from './constants';
 import { AssetBalance, BoxInfo, CoveringBoxes, TokenInfo } from './types';
 import { ErgoBoxProxy } from '@rosen-ui/wallet-api';
 import ergoExplorerClientFactory from '@rosen-clients/ergo-explorer';
@@ -12,7 +10,7 @@ import ergoExplorerClientFactory from '@rosen-clients/ergo-explorer';
  */
 export const getHeight = async (): Promise<number> => {
   const explorerClient = ergoExplorerClientFactory(
-    'https://api.ergoplatform.com',
+    'https://api.ergoplatform.com'
   );
   return Number((await explorerClient.v1.getApiV1Networkstate()).height);
 };
@@ -38,7 +36,7 @@ export const createLockBox = (
   toAddress: string,
   fromAddress: string,
   bridgeFee: bigint,
-  networkFee: bigint,
+  networkFee: bigint
 ): wasm.ErgoBoxCandidate => {
   /**
    * TODO: fix ergo native token id
@@ -48,7 +46,7 @@ export const createLockBox = (
   const lockBox = new wasm.ErgoBoxCandidateBuilder(
     wasm.BoxValue.from_i64(wasm.I64.from_str(boxErgValue.toString())),
     wasm.Contract.pay_to_address(wasm.Address.from_base58(lockAddress)),
-    height,
+    height
   );
 
   lockBox.set_register_value(
@@ -59,7 +57,7 @@ export const createLockBox = (
       Buffer.from(networkFee.toString()),
       Buffer.from(bridgeFee.toString()),
       Buffer.from(fromAddress.toString()),
-    ]),
+    ])
   );
 
   /**
@@ -69,7 +67,7 @@ export const createLockBox = (
   if (tokenId !== 'erg') {
     lockBox.add_token(
       wasm.TokenId.from_str(tokenId),
-      wasm.TokenAmount.from_i64(wasm.I64.from_str(amount.toString())),
+      wasm.TokenAmount.from_i64(wasm.I64.from_str(amount.toString()))
     );
   }
   return lockBox.build();
@@ -85,18 +83,18 @@ export const createLockBox = (
 export const createChangeBox = (
   changeAddress: string,
   height: number,
-  balance: AssetBalance,
+  balance: AssetBalance
 ): wasm.ErgoBoxCandidate => {
   const changeBox = new wasm.ErgoBoxCandidateBuilder(
     wasm.BoxValue.from_i64(wasm.I64.from_str(balance.nativeToken.toString())),
     wasm.Contract.pay_to_address(wasm.Address.from_base58(changeAddress)),
-    height,
+    height
   );
 
   balance.tokens.forEach((token) => {
     changeBox.add_token(
       wasm.TokenId.from_str(token.id),
-      wasm.TokenAmount.from_i64(wasm.I64.from_str(token.value.toString())),
+      wasm.TokenAmount.from_i64(wasm.I64.from_str(token.value.toString()))
     );
   });
   return changeBox.build();
@@ -132,18 +130,18 @@ export const getCoveringBoxes = async (
   requiredAssets: AssetBalance,
   forbiddenBoxIds: Array<string>,
   trackMap: Map<string, ErgoBoxProxy | undefined>,
-  boxIterator: Iterator<ErgoBoxProxy, undefined>,
+  boxIterator: Iterator<ErgoBoxProxy, undefined>
 ): Promise<CoveringBoxes> => {
   let uncoveredNativeToken = requiredAssets.nativeToken;
   const uncoveredTokens = requiredAssets.tokens.filter(
-    (info) => info.value > 0n,
+    (info) => info.value > 0n
   );
 
   const isRequirementRemaining = () => {
     return uncoveredTokens.length > 0 || uncoveredNativeToken > 0n;
   };
 
-  let offset = 0;
+  const offset = 0;
   const result: Array<ErgoBoxProxy> = [];
 
   // get boxes until requirements are satisfied
@@ -177,7 +175,7 @@ export const getCoveringBoxes = async (
     let isUseful = false;
     boxInfo.assets.tokens.forEach((boxToken) => {
       const tokenIndex = uncoveredTokens.findIndex(
-        (requiredToken) => requiredToken.id === boxToken.id,
+        (requiredToken) => requiredToken.id === boxToken.id
       );
       if (tokenIndex !== -1) {
         isUseful = true;
@@ -209,7 +207,7 @@ export const getCoveringBoxes = async (
  */
 export const sumAssetBalance = (
   a: AssetBalance,
-  b: AssetBalance,
+  b: AssetBalance
 ): AssetBalance => {
   // sum native token
   const nativeToken = a.nativeToken + b.nativeToken;
@@ -240,7 +238,7 @@ export const subtractAssetBalance = (
   a: AssetBalance,
   b: AssetBalance,
   minimumNativeToken = 0n,
-  allowNegativeNativeToken = false,
+  allowNegativeNativeToken = false
 ): AssetBalance => {
   // sum native token
   let nativeToken = 0n;
@@ -249,7 +247,7 @@ export const subtractAssetBalance = (
   else if (allowNegativeNativeToken) nativeToken = 0n;
   else
     throw new Error(
-      `Cannot reduce native token: [${a.nativeToken.toString()}] is less than [${b.nativeToken.toString()} + ${minimumNativeToken.toString()}]`,
+      `Cannot reduce native token: [${a.nativeToken.toString()}] is less than [${b.nativeToken.toString()} + ${minimumNativeToken.toString()}]`
     );
 
   // reduce all `b` tokens
@@ -263,7 +261,7 @@ export const subtractAssetBalance = (
         throw new Error(
           `Cannot reduce token [${token.id}]: [${tokens[
             index
-          ].value.toString()}] is less than [${token.value.toString()}]`,
+          ].value.toString()}] is less than [${token.value.toString()}]`
         );
     } else
       throw new Error(`Cannot reduce token [${token.id}]: Token not found`);