+++ apps/rosen-service/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @rosen-bridge/rosen-service
 
+## 1.0.0
+
+### Major Changes
+
+- save asset data individually for all bridged chains
+
+### Minor Changes
+
+- Upgrade to latest scanner and extractor packages
+- Add Bitcoin scanner and its observation and event trigger extractors
+- Store locked asset data in the database based on TokenMap file
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-ui/asset-calculator@1.0.0
+
 ## 0.3.0
 
 ### Minor Changes